Moved  queries/mutations from utils to auth
