+++ test/non-functional/auth/queries.ts
